more reliable _wait_until_turn
The look away part was not working since sometimes the king remains
highlighted.  It is not actually needed any more so I removed it.  I
also fixed behaviour when the king moves --- we need to update the king
position, since the initial position is actually the position before the
king has moved.