wait for *two* updates after sending a move
Without this two moves can be read at the same time by
SV_SendClientMessages so the first is never used.  The reason waiting
once is not enough is because the preceding client.move may be sent
only after the server has read messages but has not yet sent the update
for the same frame.